OC-326: split the largest PDF reporter classes and tests
RenderingSupport keeps its entry points and delegates to ReportHeaderTables and CoverageTables; PdfCellStyle.Builder and PdfBoxCanvas.PdfBoxGraphicsScope become top-level classes. TableRendererTest, PdfDocumentTest and PdfTableTest are regrouped by context into seven files.
